Fix(docs): Improve Tag docs
