+++ lib/docs/src/components/tag.stories.tsx
@@ -1,42 +1,44 @@
 import { Meta } from "@storybook/react";
 import { Tag } from "../../../core/src";
 import { Utils } from "../utils/utils";
-import { GalleryFeedbackTag } from "../../../gallery/src/feedback/tag";
+import { GalleryTag } from "../../../gallery/src/tag";
 
 const meta: Meta = {
 	title: "Components/Tag",
 	component: Tag,
 	argTypes: {
-		children: Utils.arg("text"),
-		color: Utils.arg(Object.keys(Tag.colors)),
+		children: Utils.arg(null),
+		color: Utils.arg(Tag.colors),
 	},
 };
 
 Utils.page.component(meta, {
 	primary: "sticky",
-	shots: [<GalleryFeedbackTag key="1" />],
+	shots: [<GalleryTag key="1" />],
 });
 
 export default meta;
 
 interface Props {
-	children: string;
 	color?: string;
 }
 
-export const Primary = (props: Props): JSX.Element => {
-	return (
+export const Primary = (props: Props): JSX.Element => (
+	<Tag
 		// eslint-disable-next-line
-		<Tag color={(Tag.colors as any)[props.color ?? "gray"]}>
-			{props.children ?? "Default"}
-		</Tag>
-	);
-};
+		color={(Tag.colors as any)[props.color ?? "gray"]}
+		children="Foo"
+	/>
+);
 
-export const Basic = (): JSX.Element => {
-	return <Tag color={Tag.colors.green}>Foo</Tag>;
-};
+export const Basic = (): JSX.Element => <Tag color={Tag.colors.green}>Foo</Tag>;
 
-Utils.story(Basic, { desc: `
-To begin, you need to provide a color from Tag.colors and a label via children.
-`});
+Utils.story(Basic, {
+	desc: `
+A tag requires a label, defined as a string via the \`children\` prop, and a
+color, defined via the same name prop, and should come from the \`Tag.colors\`
+list. You can try all available colors in the [All Props][1] table below.
+
+[1]: #props
+`,
+});
